requirements.txt
fastapi==0.100.0
pydantic==2.0.3
uvicorn
psycopg2-binary
python-dotenv~=1.0.0
pytest==7.4.0
pytest-mock
pytest-bdd
pytest-env
requests~=2.31.0
gitpython~=3.1.18
dpath==2.1.6
geopy==2.3.0
pandas==2.0.3
numpy==1.25.1
matplotlib==3.7.2
seaborn==0.12.2
contextily==1.3.0
haversine==2.8.0
scikit-learn==1.3.0
folium
networkx==3.1
kneed==0.8.5
hdbscan==0.8.33

dash~=2.11.1
plotly~=5.15.0
python-dateutil~=2.8.2
neo4j~=5.10.0
graphdatascience==1.7
